requirements.txt
anyio==3.7.0
async-timeout==4.0.2
click==8.1.3
dnspython==2.3.0
exceptiongroup==1.1.2
fastapi==0.96.0
h11==0.14.0
httptools==0.5.0
idna==3.4
neo4j==5.9.0
numpy==1.25.1
pandas==2.0.3
pydantic==1.10.9
pymongo==4.3.3
python-dateutil==2.8.2
python-dotenv==1.0.0
pytz==2023.3
PyYAML==6.0
redis==4.5.5
six==1.16.0
sniffio==1.3.0
starlette==0.27.0
typing_extensions==4.6.3
tzdata==2023.3
uvicorn==0.22.0
uvloop==0.17.0
watchfiles==0.19.0
websockets==11.0.3
